requirements.txt
allure-pytest==2.29.0
allure-python-commons==2.29.0
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pycparser==2.21
PySocks==1.7.1
pytest==8.0.0
pytest-cov==4.1.0
selenium==4.21.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.1
wsproto==1.2.0